api/serializers.py
from django.db import models
from general.models.customer import Customer
from rest_framework import serializers


class CustomerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Customer
        fields = '__all__'
        lookup_field = 'email'
